+++ CHANGELOG.md
@@ -19,9 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improved the throughput of the market data gathering queue by applying the rate limit per data source
 - Decreased the rate limiter duration of the market data gathering queue jobs from 4 to 3 seconds
-
-### Changed
-
+- Removed the deprecated `SymbolProfile` field from the endpoint `GET api/v1/portfolio/holding/:dataSource/:symbol`
 - Upgraded `@simplewebauthn/browser` and `@simplewebauthn/server` from version `13.2.2` to `13.3`
 
 ### Fixed